Different improvements, change rules to use floats instead of ints to allow usage of partial hours for hour-based rules.

+++ backend/src/main/java/com/planner/dto/AnnexRuleDto.java
@@ -12,5 +12,5 @@ public record AnnexRuleDto(
         Integer teacherId,
         String teacherFirstName,
         String teacherLastName,
-        Integer intValue
+        Double intValue
 ) {}
